Avoid BC break when fixing SolrQuery methods
Using the new ZPP api we can accept string|int types in strict mode.
Now we allow any string (just like before), but we also allow ints.
This way we still allow the old (wrong) way of sending in strings,
but allow those that want to do this properly, and according to the
documentation to use the correct types.